third_party/WebKit/LayoutTests/fast/css/font-face-unicode-range-monospace.html

<html>
<head>
<style>
@font-face {
  font-family: myfont;
  src: local(Courier), local('Courier New'); /* Use monospace font */
  unicode-range: U+0041; /* 'A' */
}

.test {
  font-size: 100px;
  display:inline-block;
  font-family: myfont,sans-serif;
  background: blue;
}

.sans {
  font-size: 100px;
  display:inline-block;
  font-family:sans-serif;
  background:red;
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="test" class="test">m</div>
<div id="sans" class="sans">m</div>
<script>
description('@font-face should not affect width of characters outside of its unicode-range');
shouldBe("document.getElementById('test').offsetWidth", "document.getElementById('sans').offsetWidth");
</script>
</body>
</html>
